# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервиса GophKeeper для безопасного хранения данных",
        "title": "GophKeeper API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gophkeeper.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/ping": {
            "get": {
                "description": "Проверяет доступность сервера и подключение к БД",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "monitoring"
                ],
                "summary": "Проверка здоровья сервера",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secret": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает секрет текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Создание секрета пользователя",
                "parameters": [
                    {
                        "description": "Данные секрета",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.SecretCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешное создание секрета"
                    },
                    "400": {
                        "description": "Неверный запрос"
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/secret/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех сохраненных секретов (данных) текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Получение списка секретов пользователя",
                "responses": {
                    "200": {
                        "description": "Успешное получение списка секретов",
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.SecretListResponse"
                        }
                    },
                    "204": {
                        "description": "Нет содержимого (список секретов пуст)"
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/secret/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает секрет текущего авторизованного пользователя по id секрета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Получение секрета пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID секрета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение секрета",
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.SecretResponse"
                        }
                    },
                    "204": {
                        "description": "Секрет не найден"
                    },
                    "400": {
                        "description": "Неверный запрос"
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Вход в систему с получением JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.AuthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход, токен в cookie и теле ответа",
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или пустые поля"
                    },
                    "401": {
                        "description": "Неверный логин или пароль"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Создает нового пользователя с указанным логином и паролем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.AuthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная регистрация, токен в cookie и теле ответа",
                        "schema": {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или пустые поля"
                    },
                    "409": {
                        "description": "Пользователь с таким логином уже существует"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_mdflamingo_GophKeeper_internal_model.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "github_com_mdflamingo_GophKeeper_internal_model.AuthUser": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "strongpassword123"
                }
            }
        },
        "github_com_mdflamingo_GophKeeper_internal_model.DataType": {
            "type": "string",
            "enum": [
                "TEXT",
                "CARD",
                "FILE",
                "CREDENTIALS"
            ],
            "x-enum-varnames": [
                "TEXT",
                "CARD",
                "FILE",
                "CREDENTIALS"
            ]
        },
        "github_com_mdflamingo_GophKeeper_internal_model.SecretCreateRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "data_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.DataType"
                        }
                    ],
                    "example": "CREDENTIALS"
                }
            }
        },
        "github_com_mdflamingo_GophKeeper_internal_model.SecretListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "secrets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.SecretResponse"
                    }
                }
            }
        },
        "github_com_mdflamingo_GophKeeper_internal_model.SecretResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "data_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_mdflamingo_GophKeeper_internal_model.DataType"
                        }
                    ],
                    "example": "CREDENTIALS"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "meta_data": {
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}